// src/api/__generated__/mining/model/actorDTO.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Mining API
 * 挖矿服务接口
 * OpenAPI spec version: v1.0
 */
import type { ActorDTOStatus } from './actorDTOStatus';
import type { Memo } from './memo';

/**
 * 演员信息响应
 */
export interface ActorDTO {
  /**
   * 演员名称
   * @maxLength 100
   */
  actorName?: string;
  /**
   * 演员NFT链上ID
   * @maxLength 64
   */
  actorNftId?: string;
  /** 演员IP */
  actorCollectionId?: number;
  /**
   * 等级
   * @minimum 1
   * @maximum 5
   */
  level?: number;
  /**
   * 热度值
   * @minimum 0
   */
  heat?: number;
  /** 算力 */
  computingPower?: number;
  /**
   * 挖矿系数
   * @minimum 1
   */
  miningCoefficient?: number;
  /** 派遣状态: MINING/REST */
  status?: ActorDTOStatus;
  /** 本周名义产出 */
  weeklyNominalOutput?: number;
  /**
   * 当前体力
   * @minimum 0
   * @maximum 168
   */
  stamina?: number;
  /** 头像URL */
  avatarUrl?: string;
  /** 累计完播阈值 */
  completedPlayCount?: number;
  /** 耗材数量 */
  materialCount?: number;
  memo?: Memo;
  /** 演员actorTokenId */
  actorTokenId?: number;
}